make abstract test case
